+++ cmake/ThirdPartyOCE.cmake
@@ -41,19 +41,7 @@ IF(NEKTAR_USE_MESHGEN)
     IF (OCC_FOUND)
         # We definitiely don't need to build OCE in this case.
         SET(BUILD_OCE OFF)
-
-        IF (OCC_VERSION_STRING VERSION_LESS 7.8.0)
-            LIST(APPEND OCC_LIB_LIST TKIGES TKSTEPBase TKSTEPAttr TKSTEP209 TKSTEP TKXDESTEP TKSTL TKSTEPAttr)
-        ELSE()
-            LIST(APPEND OCC_LIB_LIST TKDESTEP TKDEIGES TKDESTL)
-        ENDIF()
-
-        # Re-run to find additional components for STEP.
-        SET(FIND_OCC_QUIET OFF)
-        INCLUDE(FindOCC)
     ELSE()
-        LIST(APPEND OCC_LIB_LIST TKIGES TKSTEPBase TKSTEPAttr TKSTEP209 TKSTEP TKXDESTEP TKSTL)
-
         SET(BUILD_OCE ON)
     ENDIF()
 
@@ -67,13 +55,15 @@ IF(NEKTAR_USE_MESHGEN)
             MESSAGE(SEND_ERROR "Cannot currently use OpenCascade with Nektar++ on Windows")
         ENDIF()
 
+        LIST(APPEND OCC_LIB_LIST TKIGES TKSTEPBase TKSTEPAttr TKSTEP209 TKSTEP TKXDESTEP TKSTL)
+
         THIRDPARTY_LIBRARY(OCC_LIBRARIES SHARED ${OCC_LIB_LIST} DESCRIPTION "OpenCascade libs")
 
         UNSET(PATCH CACHE)
         FIND_PROGRAM(PATCH patch)
         IF(NOT PATCH)
             MESSAGE(FATAL_ERROR
-                "'patch' tool for modifying files not found. Cannot build boost-numpy.")
+                "'patch' tool for modifying files not found. Cannot build OCE.")
         ENDIF()
         MARK_AS_ADVANCED(PATCH)
 
@@ -93,6 +83,8 @@ IF(NEKTAR_USE_MESHGEN)
                 ${NEKTAR_EXTERNAL_PROJECT_CMAKE_GENERATOR_ARGS}
                 -DCMAKE_C_COMPILER:FILEPATH=${CMAKE_C_COMPILER}
                 -DCMAKE_CXX_COMPILER:FILEPATH=${CMAKE_CXX_COMPILER}
+                -DCMAKE_C_FLAGS="-w"
+                -DCMAKE_CXX_FLAGS="-w"
                 -DOCE_INSTALL_PREFIX:PATH=${TPDIST}
                 -DOCE_TESTING=OFF
                 -DOCE_VISUALISATION=OFF
@@ -109,6 +101,15 @@ IF(NEKTAR_USE_MESHGEN)
         SET(OCC_INCLUDE_DIR ${TPDIST}/include/oce CACHE FILEPATH "OCC include" FORCE)
         MESSAGE(STATUS "Build OpenCascade community edition: ${TPDIST}/lib")
     ELSE()
+        IF (OCC_VERSION_STRING VERSION_LESS 7.8.0)
+            LIST(APPEND OCC_LIB_LIST TKIGES TKSTEPBase TKSTEPAttr TKSTEP209 TKSTEP TKXDESTEP TKSTL TKSTEPAttr)
+        ELSE()
+            LIST(APPEND OCC_LIB_LIST TKDESTEP TKDEIGES TKDESTL)
+        ENDIF()
+
+        # Re-run to find additional components for STEP.
+        SET(FIND_OCC_QUIET OFF)
+        INCLUDE(FindOCC)
         ADD_CUSTOM_TARGET(oce-0.18.3 ALL)
     ENDIF()